Flatten repo structure to remove workspaces
