+++ examples/index.html
@@ -19,9 +19,9 @@
 
 <head>
   <meta charset="utf-8" content="width=device-width, initial-scale=1, maximum-scale=1, minimum-scale=1, user-scalable=no" name="viewport">
-  <script src="../packages/tree-finder/dist/tree-finder.js"></script>
-  <link rel="stylesheet" href="../packages/tree-finder/dist/tree-finder.css">
-  <link rel="stylesheet" href="../packages/tree-finder/dist/theme/material.css">
+  <script src="../dist/tree-finder.js"></script>
+  <link rel="stylesheet" href="../dist/tree-finder.css">
+  <link rel="stylesheet" href="../dist/theme/material.css">
 
   <style>
         body {
@@ -44,7 +44,7 @@
 <body>
 
   <script type="module">
-    import "../packages/tree-finder/dist/tree-finder.js";
+    import "../dist/tree-finder.js";
 
     const root = mockContent({
         kind: "dir",
