--- packages/tree-finder/package.json
@@ -1,50 +1,0 @@
-{
-    "name": "@tree-finder/base",
-    "version": "1.0.2",
-    "description": "browser/tree widget implemented as a custom DOM element. Built on top of regular-table",
-    "homepage": "https://github.com/tree-finder/tree-finder",
-    "bugs": {
-        "url": "https://github.com/tree-finder/tree-finder/issues"
-    },
-    "repository": {
-        "type": "git",
-        "url": "https://github.com/tree-finder/tree-finder.git"
-    },
-    "license": "BSD-3-Clause",
-    "author": "Max Klein <telamonian@users.noreply.github.com>",
-    "files": [
-        "dist/**/*",
-        "src/**/*",
-        "style/**/*.{css,eot,gif,html,jpg,json,less,png,scss,svg,ttf,woff2}",
-        "types/**/*"
-    ],
-    "main": "dist/tree-finder.js",
-    "style": "dist/tree-finder.css",
-    "types": "types/index.d.ts",
-    "scripts": {
-        "build": "tsc --build",
-        "build:watch": "tsc --build --watch",
-        "clean": "rimraf dist lib types *.tsbuildinfo",
-        "lint": "prettier --check \"src/**/*.ts\" \"tests/**/*.js\" \"style/**/*.less\" \"*.js\" \"*.json\"",
-        "fix": "prettier --write  \"src/**/*.ts\" \"tests/**/*.js\" \"style/**/*.less\" \"*.js\" \"*.json\"",
-        "postpack": "shx rm README.md LICENSE",
-        "prepack": "shx cp ../../README.md . && shx cp ../../LICENSE .",
-        "prepublishOnly": "pnpm run clean && pnpm run webpack:prod",
-        "svgo": "svgo --config svgo.yaml",
-        "svgo:uri": "svgo --config svgo.yaml --datauri=base64 -o -",
-        "test:unit": "jest",
-        "test:e2e": "playwright test",
-        "test": "npm-run-all test:*",
-        "watch": "npm-run-all -p *:watch",
-        "webpack": "webpack --color",
-        "webpack:prod": "NODE_ENV=production webpack --color",
-        "webpack:watch": "webpack --color --watch"
-    },
-    "dependencies": {
-        "regular-table": "^0.6.8",
-        "rxjs": "^7.8.1"
-    },
-    "publishConfig": {
-        "access": "public"
-    }
-}
