--- packages/tree-finder/tsconfig.json
@@ -1,11 +1,0 @@
-{
-    "$schema": "http://json.schemastore.org/tsconfig",
-    "extends": "../../tsconfigbase",
-    "compilerOptions": {
-        // "baseUrl": ".",
-        "declarationDir": "types",
-        "outDir": "lib",
-        "rootDir": "src"
-    },
-    "include": ["src/**/*"]
-}
