+++ src/element/panel.ts
