+++ src/model/clipboard.ts
