+++ style/panel/index.less
