+++ svgo.yaml
